refined cards, made them links, and moved to seperate component
