+++ client/src/global/components/index.js
@@ -1,4 +1,5 @@
 import Header from "./Header";
 import Footer from "./Footer";
+import CardLink from "./CardLink";
 
-export { Header, Footer };
+export { Header, Footer, CardLink};
